condensed shared code into a helper InputHandler class
