+++ Assets/MRTK/Providers/WindowsMixedReality/Shared/InputHandlers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 024434572f1045f4d84cc29b8996b203
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
